Move changes over to a BeanFactoryInitializationAotProcessor.
